Chances left will now show only with incorrect guess
The game runs much better this way. I replaced the times loop with a
while loop for this to work. So instead of running "chances" amount of
times, the game runs until there's a win. The only other exception is if
the player runs out of chances, which will break out of the loop and
result in a loss.